Enable better database testing
On any given test run, the list of test classes to run is determined.
If classes are provided on the command line, then those are the tests
to run. Otherwise, test classes are searched for, constrained only by
the --etl parameter to the test runner.

If any ETL tests (in TestForETL) are to be run, a test database is
created by duplicating the production database. If any other tests
(in any other dir) are to be run, a test database is created from
source and by applying any unapplied migrations.

A method modifier is provided to the test runner via executor roles
as supported in new versions of Test::Class::Moose. For each and
every test method run, this modifier makes a copy of the correct
test database for use only by one test method, sets the copy as the
current database (e.g., it will be connected to when using the
HasWebAppSchema role [which is the code's primary method of
accessing the database]), runs the test, and then drops the copied
database.

At the end of the test run, the test runner drops any of the testing
databases it created.

Because of this new process, testing can now be done without concern
of altering production data. Also, testing of transactions can be
done properly.